S15.h: extract handoff_succession to sibling module (mod.rs 870 → 658 LOC, §4.2 cap closed)
Final §4.2 Phase 2 cap. Extracts the 209-LOC handoff_succession route
handler from inference-router/src/routes/handoff/mod.rs to sibling
inference-router/src/routes/handoff/succession.rs. Body byte-identical;
only pub(super) visibility added so the routes table can still
reference it via 'use succession::handoff_succession;'.

mod.rs: 870 → 658 LOC (142 under cap of 800). All §4.2 Phase 2 caps
closed.

cargo build / clippy / test: 608 passed / 0 failed.
